test: improving tests coverage
- Testing more types formatting cases
- Covering more exceptional cases
- Improving support for graph exporting custom options

+++ tests/lib/transformers.py
@@ -111,7 +111,7 @@ def repeat(content: str, n_times: int, linebreak: bool) -> str:
 
 
 @partial_transformer
-def repeat_list(content: int, n_times: int) -> list[int]:
+def repeat_list(content: float, n_times: int) -> list[float]:
     return [content] * n_times
